requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
supabase>=1.0.3
geopandas>=0.9.0
shapely>=1.8.0
numpy>=1.20.3
pandas>=1.3.0
pyproj>=3.1.0
fiona>=1.8.20
pyogrio>=0.7.2
openai>=1.1.0
setuptools>=45.0.0
requests>=2.25.1
langchain>=0.1.17
langchain-community>=0.0.36
langchain-openai>=0.0.3
openai>=1.12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0